# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TempNest API - Temporary Email Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inboxes:
    get:
      operationId: listInboxes
      tags: [inboxes]
      summary: List user's inboxes
      responses:
        "200":
          description: List of inboxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inbox"
        "401":
          description: Unauthorized
    post:
      operationId: createInbox
      tags: [inboxes]
      summary: Create a new temporary inbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxInput"
      responses:
        "201":
          description: Created inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "400":
          description: Insufficient credits or daily limit reached
        "401":
          description: Unauthorized
  /inboxes/{inboxId}:
    get:
      operationId: getInbox
      tags: [inboxes]
      summary: Get inbox details
      parameters:
        - name: inboxId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Inbox details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "404":
          description: Not found
    delete:
      operationId: deleteInbox
      tags: [inboxes]
      summary: Delete an inbox
      parameters:
        - name: inboxId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /inboxes/{inboxId}/refresh:
    post:
      operationId: refreshInbox
      tags: [inboxes]
      summary: Refresh inbox to pull new emails
      parameters:
        - name: inboxId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Refreshed inbox with latest emails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxRefreshResult"
        "401":
          description: Unauthorized
  /inboxes/{inboxId}/emails:
    get:
      operationId: listInboxEmails
      tags: [inboxes]
      summary: List emails in an inbox
      parameters:
        - name: inboxId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
        "401":
          description: Unauthorized
  /inboxes/{inboxId}/otp-history:
    get:
      operationId: getInboxOtpHistory
      tags: [inboxes]
      summary: Get OTP extraction history for an inbox
      parameters:
        - name: inboxId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OTP history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OtpRecord"
        "401":
          description: Unauthorized

  # ── Emails ───────────────────────────────────────────────────────
...
  /subscriptions/current:
    get:
      operationId: getCurrentSubscription
      tags: [subscriptions]
      summary: Get current user subscription
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "401":
          description: Unauthorized
  /subscriptions/schedule-downgrade:
    post:
      operationId: scheduleDowngrade
      tags: [subscriptions]
      summary: Schedule a downgrade to a lower plan for the next billing period
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleDowngradeInput"
      responses:
        "200":
          description: Downgrade scheduled
          content:
            application/json:
              schema:
                type: object
                required: [success, currentPlanId, nextPlanId, planExpiryDate]
                properties:
                  success:
                    type: boolean
                  currentPlanId:
                    type: string
                  nextPlanId:
                    type: string
                  planExpiryDate:
                    type: ["string", "null"]
        "400":
          description: Invalid downgrade request
        "401":
          description: Unauthorized

  # ── Payments ─────────────────────────────────────────────────────
...
  /admin/inboxes:
    get:
      operationId: listAdminInboxes
      tags: [admin]
      summary: List all inboxes (admin)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: All inboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInboxPage"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Inbox:
      type: object
      required: [id, address, domain, isActive, createdAt, emailCount, unreadCount]
      properties:
        id:
          type: string
        address:
          type: string
        domain:
          type: string
        isActive:
          type: boolean
        isPriority:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        emailCount:
          type: integer
        unreadCount:
          type: integer
    InboxInput:
      type: object
      properties:
        customName:
          type: ["string", "null"]
        isPriority:
          type: boolean
    InboxRefreshResult:
      type: object
      required: [newEmails, inbox]
      properties:
        newEmails:
          type: integer
        inbox:
          $ref: "#/components/schemas/Inbox"

    # ── Email ────────────────────────────────────────────────────────
    Email:
      type: object
      required: [id, inboxId, from, subject, isRead, createdAt]
      properties:
        id:
          type: string
        inboxId:
          type: string
        from:
          type: string
        subject:
          type: string
        preview:
          type: string
        isRead:
          type: boolean
        hasOtp:
          type: boolean
        otpCode:
          type: ["string", "null"]
        createdAt:
          type: string
...
    OtpRecord:
      type: object
      required: [id, emailId, code, extractedAt]
      properties:
        id:
          type: integer
        emailId:
          type: string
        code:
          type: string
        length:
          type: integer
        extractedAt:
          type: string

    # ── Credits ──────────────────────────────────────────────────────
...
    Subscription:
      type: object
      required: [planId, planName, status, currentPeriodEnd, planStartDate, planExpiryDate]
      properties:
        planId:
          type: string
        planName:
          type: string
        nextPlanId:
          type: ["string", "null"]
        nextPlanName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, cancelled, past_due, free]
        currentPeriodEnd:
          type: ["string", "null"]
        planStartDate:
          type: ["string", "null"]
        planExpiryDate:
          type: ["string", "null"]
        cancelAtPeriodEnd:
          type: boolean
    ScheduleDowngradeInput:
      type: object
      required: [planId]
      properties:
        planId:
          type: string
          enum: [free, pro, business]

    # ── Payments ─────────────────────────────────────────────────────
...
    AdminInboxPage:
      type: object
      required: [inboxes, total, page, pageSize]
      properties:
        inboxes:
          type: array
          items:
            $ref: "#/components/schemas/AdminInboxItem"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    AdminInboxItem:
      type: object
      required: [id, address, userId, isActive, emailCount, createdAt]
      properties:
        id:
          type: string
        address:
          type: string
        userId:
          type: integer
        userEmail:
          type: string
        isActive:
          type: boolean
        emailCount:
          type: integer
        createdAt:
          type: string
...